chore: hydrate 변경 (club)
