+++ apps/community/src/apis/lab/remote.ts
@@ -1,9 +1,17 @@
 import { MOCK_END_POINT } from '~/constants/api';
-import type { Lab } from '~/types/lab';
 import { http } from '~/utils/http';
+
+interface Lab {
+  id: number;
+  name: string;
+  location: string;
+  site: string;
+  professor: string;
+  img?: string;
+}
 
 function getLabs() {
   return http.get<Lab[]>(MOCK_END_POINT.LABS);
 }
 
-export { getLabs };
+export { type Lab, getLabs };
