+++ apps/community/src/app/about/club/page.tsx
@@ -1,7 +1,8 @@
+import { HydrationBoundary, dehydrate } from '@tanstack/react-query';
 import { clubQueryOptions } from '~/apis/about/club/queries';
-import { Hydrate, getDehydratedQuery } from '~/utils/react-query';
 
-import { ClubCard } from '~/components/about/club/club-card';
+import { getQueryClient } from '~/utils/get-query-client';
+
 import { ClubList } from '~/components/about/club/club-list';
 import { PageHeader } from '~/components/page-header';
 
@@ -9,9 +10,9 @@ import { PageHeader } from '~/components/page-header';
 export const dynamic = 'force-dynamic';
 
 export default async function ClubPage() {
-  const { queryKey, queryFn } = clubQueryOptions.all();
-  const query = await getDehydratedQuery({ queryKey, queryFn });
-  const { data } = query.state.data ?? {};
+  const queryClient = getQueryClient();
+
+  void queryClient.prefetchQuery(clubQueryOptions.all());
 
   return (
     <>
@@ -19,13 +20,9 @@ export default async function ClubPage() {
         title="동아리 소개"
         description="경기대학교 AI컴퓨터공학부의 연구실을 소개해요."
       />
-      <Hydrate state={{ query }}>
-        <ClubList>
-          {data?.map((club) => (
-            <ClubCard key={`club-${club.name}`} {...club} />
-          ))}
-        </ClubList>
-      </Hydrate>
+      <HydrationBoundary state={dehydrate(queryClient)}>
+        <ClubList />
+      </HydrationBoundary>
     </>
   );
 }
