+++ apps/community/src/components/about/club/club-list.tsx
@@ -1,11 +1,23 @@
+'use client';
+
+import { useSuspenseQuery } from '@tanstack/react-query';
+
+import { clubQueryOptions } from '~/apis/about/club/queries';
+
+import { ClubCard } from './club-card';
+
 import * as style from './club-list.css';
 
-interface Props {
-  children: React.ReactNode;
-}
+function ClubList() {
+  const { data } = useSuspenseQuery(clubQueryOptions.all());
 
-function ClubList({ children }: Props) {
-  return <div className={style.clubList}>{children}</div>;
+  return (
+    <div className={style.clubList}>
+      {data.data.map((club) => (
+        <ClubCard key={`club-${club.name}`} {...club} />
+      ))}
+    </div>
+  );
 }
 
 export { ClubList };
